A lot of visual updates/tweaks, as well as random move rolling beta
